Partially support DWARF 5
This doesn't quite fix #14, but it gets a good chunk of the way there.
Our goals with this change were to teach magic-trace to trace code
compiled with GCC 11, which turns on DWARF 5 by default.

To that end, we implemented just the parts of the spec necessary to
provide magic-trace symbol and file names. We're not sure if anything
else in DWARF 5 works, but at least it seems not to crash on the ELF
files we tested.

Just to point out one example of something still not supported, owee
still doesn't support split debug info (aka .dwo files). We'd like
it to one day, but it sounds harder than what we've done here.

This patch was authored jointly by @billduff and @cgaebel.

Testing: We tested this by running magic-trace on itself and seeing
symbols and filenames from C functions compiled with GCC 11.